project: Change project structure
Instead of having a single script add a pyhcam package with
margin_analysis.py module and a phycam-margin-analysis entry point to
access the modules main function.

This change makes the code reusable for other modules and with the
console_scripts entry point it is now possible to call the modules main
function easily from the cli.

Also while at it don't add an additional src package but rather ony look
for packages inside the src directory.

Signed-off-by: Stefan Riedmueller <s.riedmueller@phytec.de>